Started graphs
